refactor(managed-agents): split managed-invocation god-file into focused modules
Break packages/runtime/src/agents/managed-invocation/index.ts (3,531 lines,
one 2,468-line invocation-service god-class plus 6 lease-manager classes)
into 18 sibling modules, mirroring the message-pipeline/ split (24139b8b):
one file per lease manager, the invocation service decomposed into named
phases (admission guard, authority guard, lease lifecycle, lifecycle events,
prompt delivery, recovery checkpoint, records, environment), with
invocation-service.ts as a thin coordinator. index.ts remains a pure barrel
preserving the exact prior public surface.

Behavior-preserving. Rebased onto 59e73ea5 (managed economic lifecycle
session events) with no logic conflicts — that commit only touched sibling
files and additive barrel exports.

Resolves #54.
