+++ packages/runtime/src/agents/managed-invocation/lease-errors.ts
@@ -1,0 +1,12 @@
+import { ManagedAgentRuntimeAdmissionError } from "./errors.js";
+
+export class ManagedAgentLeaseAcquireError extends ManagedAgentRuntimeAdmissionError {
+  constructor(
+    message: string,
+    readonly sideEffected: boolean,
+  ) {
+    super(message);
+  }
+}
+
+export class ManagedAgentWorktreeReviewRequiredError extends ManagedAgentRuntimeAdmissionError {}
